+++ src/lib/builtins/null.ts
@@ -1,0 +1,7 @@
+/**
+ * @module kryo/builtins/null
+ */
+
+import { NullType } from "../types/null";
+
+export const $Null: NullType = new NullType();