+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Fix]** Fix `TsEnumType` name.
+
 # 0.8.0 (2018-12-03)
 
 - **[Breaking change]** Change `variantRead` result from `[variant, value]` to `{variant, value}`.